Split into index building class and matching class
